+++ src/modules/todos/todos.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { TodosService } from '../service/todos.service';
-import { TodosResolver } from '../resolver/todos.resolver';
+import { TodosService } from '../todos/services/todos.service';
+import { TodosResolver } from '../todos/resolvers/todos.resolver';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { TodosEntity } from '@entities/todos.entity';
-import { TodosRepository } from '../repository/todos.repository';
+import { TodosRepository } from '../todos/repositories/todos.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([TodosEntity])],
